requirements.txt
absl-py==0.8.1
setuptools==40.3.0
backports.weakref==1.0.post1
bleach==1.5.0
Click==7.0
enum34==1.1.6
Flask==1.1.1
funcsigs==1.0.2
gunicorn==19.9.0
h5py==2.10.0
html5lib==0.9999999
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.3.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.5
Pillow==6.2.1
protobuf==3.10.0
PyYAML==5.1.2
scipy==0.19.1
six==1.12.0
tensorflow==1.13.1
Werkzeug==0.16.0
DateTime==4.3